terapia_visual_adapter

Module notifier

Adaptadores de Notificaciones

Este módulo contiene los adaptadores que implementan el puerto terapia_visual_domain::ports::SystemNotifier para diferentes plataformas.

Actualmente solo se proporciona la implementación para Tauri, que utiliza las capacidades nativas del sistema operativo a través de Tauri.

Adaptadores disponibles

Adaptador	Plataforma	Tecnología

TauriSystemNotifier	Windows, Linux, macOS	Tauri + tauri-plugin-notification

Uso

use terapia_visual_adapter::notifier::TauriSystemNotifier;
use terapia_visual_domain::ports::SystemNotifier;
use tauri::AppHandle;

let notifier = TauriSystemNotifier::new(app_handle, icon_inactive, icon_active);

Re-exports

pub use tauri_notifier::TauriSystemNotifier;

Modules

tauri_notifier

Adaptador de Notificaciones para Tauri